Fixed: bad bug in RSTC.BAT.Base.bestAccel
ntgDiff (aka quality) was called for the future observation, but for the
current situation instead of the future situation (which incorporates
the tested acceleration and the wait for the future observation)!

With the bug, bestAccel always returned NaN, because the functions
to be optimized (the quality aka ntgDiff) just were constant!

To make bestAccel a little clearer, I also rewrote some of the helpers.

This bug-fix has important runtime implications:
logs/overtake-0002.log with RSTC.BAT.Regression: 39 s
logs/overtake-0002.log with RSTC.BAT.Progression: 19 s (real time!)
That's great :-).