Found semantics problem: order of interleavings.
The order of the interleavings plays a much more important role in our
plan recognition examples (unfortunately, because it's an internal thing
one really shouldn't depend on).

Now the order of the interleavings of the new semantics is the same (or
at least very similar) as in the old semantics.
And suddenly the tests work again.

I think there may be some justification for exactly this ordering, if
the match-actions are the left-most program. We should have a very close
look on this.