Golog's Conf is now just the current Tree (Node a b).
Before we wrapped around the current Tree (Node a b) the Conf type which
additionally held the current situation.
We now start with a tree which is rooted in Val (Node s0 ...), and all
calls to trans advance to the next Val node. This Val node always
contains the current situation.

This simplifies the code in some places a little bit.
The restriction that treeND, treeDT, and trans all return a Val-rooted
tree is not enforced in the code. I tried a phantom type for that, but
that probably introduces more confusion. Also, I'm not sure whether the
Monoid instance would work with this phantom type.