Golog.Interpreter: fixed sync loop in test actions.
Assume a configuration c' was reached from a configuration c by
executing a successful test action.
The old exec function just copies the configuration c to c'. This
particularly copies the payload. (This was something we had thought
about earlier, because for that reason we couldn't increase d to d+1 in
the DT interpreter.) The IO interpreters have as payload an action which
returns a configuration, and in the described scenario this
configuration proceeds once again with the test action, because it's
just a copy of c!

This is clearly unwanted behavior as it leads to loops.

Unfortunately the code doesn't get nicer due to the additional functions
that update the payload in case a test action was executed.

A reasonable consideration seems to be to drop test actions. They can be
simulated by a primitive action `Test (\s -> ...)' whose precondition
evaluates the formula.
To avoid cluttering the situation term, the do_ function could ignore
it: do_ (Test _) s = s.

Interestingly the Test construct only occurs in two places in the whole
Golog.Interpreter module: the data definition of Atom and the case of
the exec function.