Golog.Interpreter: fixed wrong resolution of nondeterminism.
When a program p* where p has a negative reward is executed, the best
execution is to stop immediately, i.e., the program should be final in
treeDT.

The program leads to the tree
  Val S0
    Alt
      Empty
      Val (Do p S0)
        Alt
        ...

The old resolve function went from the root Val node to the Alt node and
there compared all the alternatives. The chooseDT function tells the
best function to treat Empty same as Val Flop nodes, i.e., failed
execution, and returns the worst possible reward (-1/0,minBound).

The new resolve function hands over the reward of the Val S0 node, more
generally, the latest Val node's reward, to the resolution function.
Thus when best encounters an Empty node, it can return this reward.