TORCS: working on simulation of CarState.
Fixed a few mistakes in the simulation of the state for offline
execution.

We want to use treeDTIO, but for that we need a reasonable reward
function. For example, to steer the car to the middle, the reward should
take into account the difference between the current trackPos and the
goal trackPos.

For that we need to be able to predict the effect of steering.
That's a problem still to be solved. Learning?