TORCS: learned steering/accel/brake influence.
Wrote a Python script that uses scikit-learn to from our examples.
The examples were recorded from a berniw car driving a bunch of
tracks. There are two datasets, one for training and one for validation.

Used SVR for learning. Reason: SVR was the first regression method I
found in the scikit-learn documentation.

The SVR prediction (at least in scikit-learn) is quite slow. For that
reason we generate lookup-tables. This also is an easy solution to bring
the scikit-learn results from Python to Haskell. More precisely, the
Python script generates simple C code which is interfaced by the module
TORCS.LUT.

The C code generated by the Python script is actually ugly. It would be
better it generated a large array literal instead of the initialization
code. For now I converted the code by hand (thanks to vim).